moved files in base path to linter folder
