+++ src/io/base.js
@@ -1,6 +1,6 @@
 import { oneLine } from 'common-tags';
 
-import { FLAGGED_FILE_MAGIC_NUMBERS_LENGTH, MAX_FILE_SIZE_MB } from 'const';
+import { FLAGGED_FILE_MAGIC_NUMBERS_LENGTH, MAX_FILE_SIZE_MB } from 'linter/const';
 
 /*
  * Base class for io operations for both an Xpi or
