+++ src/linter/libraries.js
