+++ src/linter/miner_blocklist.js
