+++ src/rules/javascript/webextension-api-compat-android.js
@@ -1,7 +1,7 @@
 import bcd from 'mdn-browser-compat-data';
 
 import { ANDROID_INCOMPATIBLE_API } from 'messages/javascript';
-import { createCompatibilityRule } from 'utils';
+import { createCompatibilityRule } from 'linter/utils';
 import { hasBrowserApi } from 'schema/browser-apis';
 
 export default {
