+++ src/rules/javascript/webextension-api.js
@@ -1,5 +1,5 @@
 import { isDeprecatedApi, isTemporaryApi } from 'schema/browser-apis';
-import { apiToMessage, isBrowserNamespace } from 'utils';
+import { apiToMessage, isBrowserNamespace } from 'linter/utils';
 
 export default {
   create(context) {
