+++ src/scanners/filename.js
@@ -2,7 +2,7 @@ import { extname } from 'path';
 
 import BaseScanner from 'scanners/base';
 import * as messages from 'messages';
-import * as constants from 'const';
+import * as constants from 'linter/const';
 
 export default class FilenameScanner extends BaseScanner {
   static get scannerName() {
