+++ tests/unit/test.collector.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-named-default
-import { default as Collector } from 'collector';
+import { default as Collector } from 'linter/collector';
 
 import { fakeMessageData } from './helpers';
