+++ tests/unit/test.linter.js
@@ -3,7 +3,7 @@ import fs from 'fs';
 import { oneLine } from 'common-tags';
 
 import Linter from 'linter';
-import * as constants from 'const';
+import * as constants from 'linter/const';
 import * as messages from 'messages';
 import ManifestJSONParser from 'parsers/manifestjson';
 import BinaryScanner from 'scanners/binary';
